apps/catdog_blinker/requirements.txt
absl-py==1.4.0
array-record==0.4.0
astunparse==1.6.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
contourpy==1.1.1
cycler==0.12.1
dm-tree==0.1.8
etils==1.3.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.55.2
gast==0.4.0
google-auth==2.36.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.68.1
h5py==3.11.0
hexformat==0.2
hidapi==0.10.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
keras==2.13.1
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.2
pandas==2.0.3
pillow==10.4.0
progress==1.5
promise==2.3
protobuf==3.20.3
psutil==6.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.1.4
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2024.2
pyusb==1.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-datasets==4.9.2
tensorflow-estimator==2.13.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-metadata==1.14.0
termcolor==2.4.0
tf-keras==2.15.0
tf-slim==1.1.0
threadpoolctl==3.5.0
toml==0.10.2
tqdm==4.67.1
typing_extensions==4.5.0
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.6
wrapt==1.17.0
zipp==3.20.2
